Clear credentials when disabling lock pattern
`locksettings set-disabled` doesn't have an effect if the device has
credentials set. This shows up as an issue on devices that don't support
insecure locks where we always set a credential as part of the
`LockScreenSession` setup. This CL clears the credentials when disabling
the lock screen if we're on such a device. Additionally the `--old`
param is removed from `locksettings set-disabled` as it has no effect.

Bug: 298437313
Test: atest CtsWindowManagerDeviceActivity:android.server.wm.activity.ActivityVisibilityTests
Change-Id: I2bd2785d2577d532d6525b3e27bfbe8aa366ae21