Move RootTargetPreparer after DeviceSetup to run tests as shell user
The testRegisterReceiver_notExported_broadcastNotReceived registers
a receiver with the RECEIVER_NOT_EXPORTED flag, then uses a shell
command to have a different UID attempt to send a broadcast to the
receiver; since only the root and system UIDs can send a broadcast
to an unexported receiver, the broadcast sent by the shell user
should not be delivered. However, the DeviceSetup preparer was added
after the RootTargetPreparer in the test config; this preparer will
undo the adb unroot command that was run previously, so when the
broadcast is sent with the shell command, it is sent under the root
UID which is delivered to the receiver and causes the test to fail.
This commit moves the RootTargetPreparer after the DeviceSetup to
ensure the broadcast commands runs as the shell user.

Fixes: 294281377
Test: atest ContextWrapperTest
Merged-In: I34677b371dcde64231d32d3c213d2d6c4ff35656
Change-Id: I15035ca4ef374eec726962588f930cc9ca1a3142