Add test for fallback IMM parameter drop behaviour
When issuing an IME show request on the incorrect display, and going
through the fallback IMM instance onto the right display, the incorrect
function call was being used, resulting in the statsToken and reason for
the request being dropped.

This was fixed in [1], with the test for this behaviour added now.

  [1]: Ib6b1c0c1822be5de6b246a6c662d1d93ff4a0331

Test: atest MultiDisplayImeTests#testFallbackImmMaintainsParameters
Bug: 294992637
Change-Id: Ibb1b05350b4a96420e9d15618e75faf6fbe26710