Add checks to only test frame rate override with divisors
Game frame rate override previously only considers a small range
of display refresh rates such as 60, 90, 120Hz. As more and more
kinds of display is in practice, we need to update this test to
check if we are overriding the frame rate that is the divisor of
the display refresh rate.

Bug: 296988493

Tested on Pixel 7 Pro userdeug
Test: atest CtsFrameRateOverrideTestCases

Change-Id: I1ef864a361a4b1c332ba4330ef14feaadfd9e9cb