Wait for vold resets on HSUM configs in appdataisolation test
AppDataIsolationTests#testDirectBootModeWorks is intended to verify
Direct Boot mode functions as expected by only granting access to
device encrypted storage before the user has unlocked the device.
Part of the test verifies that credential encrypted storage is
available after the device is unlocked, but on HSUM configs, vold
has been observed to reset several times after the device is
unlocked; this causes flaky behavior as the test can fail if it
attempts to verify CE storage in the middle of a vold reset. This
commit attempts to resolve this by waiting for vold to stablize by
querying for the external and obb directories and not proceeding
until valid values are returned for both. This may not completely
resolve all test failures if the query occurs right before vold
resets.

Bug: 299426566
Test: atest AppDataIsolationTests#testDirectBootModeWorks
Change-Id: Idccdf533e4e1e1098c68e9db04c97b267e1df80f
Merged-In: Idccdf533e4e1e1098c68e9db04c97b267e1df80f