Delete external files after appdataisolation test
AppDataIsolationTests#testDirectBootModeWorks is intended to verify
Direct Boot mode functions as expected by only granting access to
device encrypted storage before the user has unlocked the device;
part of the setup for this test will verify that the files created
by the test do not yet exist on the device. However, in a HSUM
config, this test was observed to leave the external and obb files
on the device when a subsequent invocation was attempted. The file
in the external directory was still owned by the previous UID of
test app while the file in the obb directory was reported to be
owned by the new UID of the app. This test deletes these external
files after a test invocation to ensure the environment is clean
for a subsequent execution.

Bug: 299426566
Test: atest AppDataIsolationTests#testDirectBootModeWorks
(cherry picked from https://googleplex-android-review.googlesource.com/q/commit:c39a8b195b31586ccbf3fe73ea4044aeb0ac37f0)
Change-Id: Ifdd2674620b61c832187160c2fbe870d48f4766e