Run specific ComponentName tests for the sandbox
Some tests in ComponentNameTest tests with a ComponentName created using
the test context and current test class, and then verifies the package
and class.

In the case of the SDK sandbox, when a ComponentName is created using
its context, the package name will always be something like
"com.android.sdksandbox". However since SDKs cannot declare components,
and therefore no ComponentName specific to SDK components will be used,
CTS tests that check this kind of behaviour are not needed. The system
can still declare components for the sandbox and use it, but the package
name is tested separately in the AdServices module and CTS tests are not
required. Since these specific failing tests are not applicable to the
sandbox, it can be avoided.

Tests that check the ComponentName declared in a different package (or
other behaviour) are retained for the sandbox.

Bug: 300461475
Test: atest ComponentNameTest -- --enable-optional-parametrization
--enable-parametrized-modules --module-parameter run_on_sdk_sandbox

Change-Id: Ie6de6341c937e93cc0f9ab3561fcb39c28a34e78