Fix MultiDisplayActivityLaunchTests for Auto
Some tests launches an Activity on top of Home and move it to 2nd
display. And it checks if the moved Activity becomes the top focused
Activity.
But, in Auto, CarLauncher has a TaskView, so when the test Activity is
moved and CarLauncher become visible, CarLauncher launches the embeded
Activity and which prevents the test Activity from being top focused.
This CL makes Home visibile in advance for Auto devices, so the test
Activity can't be hindered by launching the embedded Activity.

Bug: 301387772
Test: atest MultiDisplayActivityLaunchTests
Change-Id: I26e826d88fddac075de27c3790e55d2f1b4ad611