Refactor for CTS test by changing inheritance
Reference to test service should be set at onBind and cleared at
onUnbind and not cleared at every setUp/tearDown since onBind is only
called for the first client connecting to the service.

Reference to QuickAccessWalletDelegateTargetActivityService should not
be stored by TestQuickAccessWalletService. Otherwise,
registerWalletServiceEventListener is done in
TestQuickAccessWalletService but sendEvent will use the reference that
was created by QuickAccessWalletDelegateTargetActivityService for
'testAddListener_sendEvent_success' when the whole test module is
executed. This leads to that onWalletServiceEvent is not sent correctly
and the test time out.

We solve this by refactoring the inheritance so both
TestQuickAccessWalletService.java and
QuickAccessWalletDelegateTargetActivityService.java extends a new
service class called TestBaseQuickAccessWalletService.java which in turn
extends the QuickAccessWalletService. This way each of our services can
have their own onBind and onUnbind methods and set/clear their weak
references as needed without interfering with the other. Shared code is
placed in the new super class TestBaseQuickAccessWalletService.java

Bug: 302691915
Test: atest CtsQuickAccessWalletTestCases
Test: run cts -m CtsQuickAccessWalletTestCases
Change-Id: I1bd62bfbdd0b44f9fc9704e9f9b34315cd3592f9