Make JvmtiHostTestAllocationTracking more deterministic
This tests global and per thread allocation tracking. When global
allocation tracking is enabled it is not possible to deterministically
check all allocation events across all the threads. This CL modifies it
so that we only check for events from the current thread and ignore any
other events. This also adds a check that we only receive the events
from the expected thread when per-thread allocation tracking is enabled.

Test: atest CtsJvmtiTrackingHostTestCases android.jvmti.cts.JvmtiHostTestAllocationTracking#testJvmti
Bug: 257921456
Change-Id: Icfacfd6ec711c79ba1c75bea03f4a8450bb90030