Fix flaky cts WallpapermanagerTest
testAlwaysUpdateWallpaperPermission_allowOutOfFocusWallpaperCommand
is observed to be flaky. After some investigation, we have learned
that we are not using ShellPermission correctly and we also did not
allow enough time for the wallpaper command to be passed along the
IPC stack.

To confirm the effectiveness of the change, the tests is ran for
100 iterations and has observed no failure.

Bug: 298201805
Bug: 294712211
Bug: 297892873
Test: atest .. --iteration 100
Change-Id: I81b0ca8201d8f4372404639f5c64b10e3ad8a6f7