Reduce flakiness of DisplayShapeTests
The previous test may trigger display configuration change
after leaving. If the device is a bit slow, the next test
activity may be delayed or being covered by a screenshot layer
for display change, which will also delay the activity window
to gain focus. So wait transition to finish at the previous
test class, and increase wait time for focus.

Bug: 304486679
Test: DisplayShapeTests
Change-Id: Ic06848b83e1930dc9b28c59cd81d9b9c4ea86e9b