Requires cts NotificationManagerTest aconfigflag
Sets FLAG_RANKING_UPDATE_ASHMEM aconfig flag to be required for
testRankingUpdateSentWithPressure. This test used to rely on a Sysprop
flag. We're moving to using an aconfig flag in the underlying code.

We don't manually set the flag because it'll be flipped in staging.

Bug: 284297289
Test: Manually set flag, Manually run atest
NotificationManagerTest#rankingUpdateSentWithPressure on local device,
ensure the flag is being set and the test passes.

Change-Id: I160734c2a11bbc345e36f0a462c61eb20c2f56f9