[SettingsProvider] ignore unexpected exceptions in testInvalidRingtoneUriIsRejected
Some OEMs choose to throw a security exception if a cross-user ringtone
file is set in SettingsProvider. The AOSP implementation ignores such
failures quietly, which doesn't throw an exception. Since the purpose of
the test is to verify that such invalid, cross-user ringtone cannot be
set, it doesn't matter if an exception was thrown or not. Modifying the
test to be more general would allow partners to have their own
implementation of exceptions without giving an indicative of a
vulnerability.

Test: atest android.provider.cts.settings.Settings_SystemTest#testInvalidRingtoneUriIsRejected
BUG: 307480718

Change-Id: Idf852462fc79fffa5745b6b131b85e72a3e1af4e