DropBoxTests for legacy behavior
This CL adds a new module to keep testing both the old behavior (before V) and the new behavior (V and after).
The Dropbox behavior is based on 2 factors: SDK version and the feature flag `enable_read_dropbox_permission`. The feature flag is enabled all the time in the test to check if Dropbox always behaves correctly on different target SDK versions.

Test: atest DropBoxTests
Bug: 307697406
Change-Id: If2a803686a4c244bfd877437062b06d20a322ff4