Fix MediaProjection CTS for watch devices
CTS tests for watches were broken because the new logic introduced caused them to try and look for a projection mode spinner, when its not present in the watch's dialog. Did some mild refactoring to fix the test logic for watches to now just scroll to the recording button and start.

Fix: 308969624
Test: atest CtsMediaProjectionTestCases
Change-Id: I224c8405de1935835c25865a3f13bd607f177f3d