[Autofill Framework CTS] Add rule to not run CTS test when device is in half-folded state.
Previously in @Before command, we have a line to disable running test in
half-folded state: http://shortn/_lC8PYuk3rj. However as stated in this
public doc:
https://developer.android.com/training/testing/instrumented-tests/androidx-test-libraries/rules#activityscenariorule.
The @Rule statement will actually run before @Before.

This poses no issue when the device is in portrait mode because the problematic line http://shortn/_KxRgtYCrL6 running in @Rule is a no-op when device is in portrait. However, in landscape mode, the line would cause the test to fail.

Adding a check before the problematic line so the problematic
line won't execute. The original line in @Before then would skip tests running in half-opened mode.

Bug: b/297413382
Test: atest CtsAutoFillServiceTestCases
Change-Id: Ia2e991ad0c35425ac3d5fd5825459a9f3bd4226b