BluetoothA2dpTest: Migrate to AndroidJUnit4
The current assertThrows implementation is wrong as it
does not make the test fail if the Runnable didn't throw
any exception.

test_setOptionalCodecsEnabled was relying on this behavior.
setOptionalCodecs binder call is oneway and exception canno't
propagate, so permission issues are silently swalowed instead
of reported to the caller. To do the migration assertThrows
as been remove from those test to keep the previous behavior.

Test: atest BluetoothA2dpTest
Change-Id: I0a4f1964221742f8b86e94e048e494018ca81081