Provide activity to VirtualDisplayActivityScenarioRule
In this refactor, we provide the activity class to the virtual display
activity rule.
This allows us to launch other activities inside virtual display. For
example, we could potentially have an activity that creates two separate
windows. We can then have per-window input event getters in that
activity.

This refactor is needed in order to add CTS tests for multi-device input
feature.

Also in this CL, kt formatting issues are fixed.

Bug: 211379801
Test: atest CtsInputTestCases
Change-Id: I11e009413c983a86155a257695e1aaf801846c78