Fixes the test long executed time
The test was used to verify the partial activity transition, but
it was no longer reasonable after commit 79b6065, where launching
activities into split-screen is now an atomic API. So, let's
wait for the ending state.

Meanwhile, the primary split-screen task is not longer docked and
therefore it won't be paused. The expected sequences should be updated
as well.

Bug: 311609690
Bug: 311609088
Test: atest ActivityLifecycleLegacySplitScreenTests
Change-Id: I754156efc4f61730f857352c8d178303caf742a8