Relax timeout waiting for testing result
Currently we have 8 seconds of timeout waiting for various of testing
result. This timeout is fine when testing locally, but it can cause
problems when running on the testing harness, where the device may have
very high CPU load.

We remove the timeout to wait indefinitely when the method signature
allows. If the API enforces timeout, we choose a very large value as 5
minutes.

Bug: 308436300
Change-Id: Ic01eedf288941a0ac808fdaea4edd2c457a18e6a