Increase the number of pointers in MotionEventTest
Before this CL, the test assumed that 17 pointers is too many for a
MotionEvent. However, there is recently a lot of interest in increasing
the number of pointers supported on Android.

To keep the test working for Android implementations where a large
number of pointers is supported, increase the value in this test from 17
to 117.

Bug: 285445200
Test: atest android.view.cts.MotionEventTest
Change-Id: I849333d6697af06b32aba0ed79e6aa6880f44449