Update AppKilledTest based on running in background thread behaviour
When MainContentCaptureSession is running in the main thread, the RESUME
session will not be able to be added in this test. Because it will be
blocked in the main thread as there are other works scheduled on the
main thread in the test. However, this is no longer the case with the
work being moved to the background thread.

The primary goal of this test is to ensure that the session would be
finished by the time the process is killed (ag/7077485). This is still
the case with running in the background thread. The differece now is we
don't make assumption that RESUME session won't appeared. Instead, we
check view level events won't be added.

Bug: 313438684
Test: AppKilledTest
Change-Id: Ia90cb5a33c11a265122d6f78b33cc3446d87392c