CtsWindowManagerDeviceTestCases: Test fails when Toast is moved to top of display
When the toast moved to top of display, the touch is not received by the
activity content area as the action bar is present behind the toast but not the activity content.

Hiding the status bar during this test, so that the activity area will move up and the test pass.

Bug: 312839622
Test: Executed the CtsWindowManagerDeviceTestCases module test cases
Change-Id: I40bc6c95191cd44017744f10b09bdb17a5cf5877