Fix flakey wifi coex test
testListenOnCoexUnsafeChannels tests that setCoexUnsafeChannels results
in CoexCallback being updated with the given channels, which only
happens if the default algorithm is disabled.

Unfortunately there's no API to detect if the algorithm is enabled or
not, so there's a possibility that the default algorithm updated
CoexCallback in the middle of the test, causing a mismatched callback to
happen after setCoexUnsafeChannels.

To avoid mistaking the mismatched callback for one caused by the setter,
set the channels 5 times and verify 5 callbacks to have better
confidence that the callbacks are caused by the setter.

Bug: 314392808
Test: atest android.net.wifi.cts.WifiManagerTest
Change-Id: I3482db96e8abc13570b29a4e2d4dc4329e778706