Fix BackgroundActivityLaunchTest failure with "next" config
The main branch "next" config restricts client target sdk ver to 34,
whereas in trun_staging there is no such restriction - with The Background
Activity Test App, it is "current" = 1000. ASM is enabled in main branch
but with backward compatibility - not enabled for target sdk less than 35 (V).
So the test fails for "next" config. Put a fix to assume target skd is >= 35
before running the test.

Bug: 308158793
Test: BackgroundActivityLaunchTest#testBackgroundActivityBlockedWhenForegroundActivityNotTop
Change-Id: I993235ef160dcb76b0051d2f4ff2039b4f0805cd