Skip AnnotationTest.testAnnotation
This test in its current form requires updates to a dynamic_config file every time annotated system APIs are added in main for mainline modules. As we cannot update the dynamic config any time there is a change, it is better to skip until we redesign it to be more useful.

Bug: 301075649
Change-Id: Ic615d45ccd41be673e40e553aeb7433609034232
Fixes: 318779775
Test: TH