Work around ARC-specific EXDEV from/to Download in testRenameDirectory
The ScopedStorageDeviceTest#testRenameDirectory test checks if a
non-empty media directory can be moved from/to the Download directory
to/from the other parts of /storage/emulated/0/ using File.renameTo().
This rename fails with EXDEV in ARCVM, where
/storage/emulated/0/Download resides in a different device than the rest
of the primary emulated volume. Therefore, this CL modifies these checks
in testRenameDirectory so that when the rename fails, the test falls
back to recursively copying the directory with Files.copy() and deletes
the source.

Bug: 304370986
Test: atest ScopedStorageDeviceTest#testRenameDirectory passes
  on Android phone (tested on lynx) and ARC (betty-arc-v).

Change-Id: I271aa50c7ab6ee0c57e60d1fbad9ba5fca5f4300