Use the dimensions of the provided display for Uinput touch device
Before this CL, we provided a way for the caller to give a custom size
to the "register" command. However, all of the callers ended up just
calling the function with the display size. Moreover, different callers
used different approaches for finding the display size.

In this CL, simplify this by converting the parameter to bool:
true - use the dimensions of the display when creating the touchscreen
false - use the dimensions specified inside the json file

Another small change in this CL is to make the companion object public.
This will allow kotlin callers to dump the input state to logcat.
Previously, it was only possible for java callers because the companion
object was private.

And another small change in this CL is to fix the formatting in some of
the files, per the recommendations from the linter.

Bug: 294553757
Test: atest CtsInputTestCases CtsInputHostTestCases
Change-Id: Id3d6f791a2d7333a2eb03cb660d9ef5123399385