Wait for window info on the specified display
Before this CL, the "waitForWindowInfo" function only worked for the
default display.

In this CL, we generalize it to support other displays. Specifically,
this is useful for virtual displays used in the input tests.

Furthermore, when a view is provided, we can find out what display it's
supposed to be on, and then use the display as the target to make sure
it becomes visible.

This is useful for determining when input can be injected to the
window.

Bug: 294553757
Test: atest CtsInputTestCases
Test: atest WindowInputTests
Test: atest SurfaceControlViewHostTests
Change-Id: If325bf066d97c1eec135674edd9546039aecba09