Use matchers for checking motion event properties
These are more flexible, because they allow the caller to specify
precisely the types of checks that should be performed on the
MotionEvent. Not all callers care about the same properties that are
set on the event.

This also brings the usage of the kt code closer in line with the native
tests, which are also matcher-based. In this case though, there aren't
too many current uses, so we can take the opportunity to force this new
behaviour onto everyone.

Bug: 294553757
Test: atest CtsInputTestCases
Change-Id: I2ab6aae6a73030ddcef42a9559ab7742bab5ea78