[cts/PreVerfiedDomains] tests for Session APIs
This CL tests the basic setter and getter APIs for preVerifiedDomains.

To test the setPreVerifiedDomain API, we first need to change the
default instant-app-installer on device. That can only be done on ENG
builds by disabling the default instant-app-installer and enabling the
temporary activity defined in the test as the new instant-app-installer.

BUG: 307327678
Test: atest InstallPreVerifiedDomainsTest

Change-Id: I5317dfe8090e98b0506fe216dd96e6f6934946b0