[DON'T BLOCK] Test ownership migration rules
This CL is created as a best effort to migrate test targets
to the new android ownership model. If you find incorrect or unnecessary
attribution in this CL, please create a separate CL to fix that.

For more details please refer to the link below,
go/new-android-ownership-model

Ignore-AOSP-First: cherry picked from aosp/2951185
Bug: 304529413
Test: N/A
Change-Id: I13ce828fdda44c5bcad9ccab8a849b714c0bef08