CTS tests for ComponentCaller#getUid/Package in onNewIntent
These tests test the new Activity#getCurrentCaller and
Activity#onNewIntent(Intent, ComponentCaller), specifically for
ComponentCaller#getUid/getPackage APIs.

The test class has 2 new singleInstance activities that are launched
twice to ensure the #onNewIntent method is reached. Once it's reached,
the test checks if the caller identities were meant to be shared.

To ensure the new intent caller is implemented correctly, the new tests
do the initial launch with the opposite sharing to the intended one for
the new intent case. For example, if the new intent caller's identity is
not meant to be shared, then the initial call is launched with the
identity shared.

Bug: 293467489
Test: atest CtsAndroidAppTestCases:android.app.cts.ShareIdentityTest
Change-Id: I59591f926237758e1cf72f642445e1adcf2877dc