Make e-call dropbox tests more resilient.
Looking at the CTS failure, it looks like one of the local logs in Telecom
wrapped between when the dropbox entry was captured and when the test
took the telecom dumpsys.  This resulted in the expected data not being
present.

Changing so that we need ~50% of the expected (20) lines in the realtime
dumpsys to be in the dropbox entry.  Realistically that's more than
enough and should make this test more reliable.

Fixes: 324507678
Test: Run CTS
Change-Id: I1d1d384dc27f2403ff7366494308279d2e5fad69