Test that wearable hotword does not trigger AppOpsManager noteOp
Ideally, we want to test that the mic indicator is off, but there does
not seem to be an API for CTS to verify that, so we just verify that
AppOpsManager.noteOp is not called for the record audio op

Also disabling this test on watches since WearOS does not support
WearableSensingManagerService. The other wearable tests will be disabled
for watches in another CL.

Bug: 310055381
Test: Ran the new test, will rely on presubmit for the existing CTS
tests since some of them are flaky.

Change-Id: I419402552b5373a63a35bdc3db59b66d2c314192