Add KeyboardConfigured and KeyboardSystemsEvent tests
These atoms are included as part of Android 15 CDD and we
need to have basic tests, testing the metrics pipeline for
keyboard atoms.
We are keeping the restrictions imposed by these tests to a minimum
for now. If we decide we want to impose requirements on OEMs to have
certain shortcuts, etc, then we can add more CTS tests enforcing those
shortcuts. As of now the purpose of these tests is to ensure the metrics
pipeline to log keyboard atoms is intact and working.

For Keyboard configured atom, we just test if keyboard configured
atom is logged if any keyboard is connected.
For Keyboard systems event reported, test if `CAPS_LOCK` key events
are logged, which we feel is something that won't change across OEMs.

Bug: 292090789
Test: atest InputAtomsTest
Change-Id: Icd229d4ae0fe1f5f586f8a3800dd577c7be992d6