Offer baseline `RavenwoodFlagsValueProvider`.
Many test authors are starting to use `CheckFlagsRule` across suites
like CTS, so instead of recommending that they use `null` when on
Ravenwood, offer them a clear behavior that is either "all-on" or
"all-off".

Eventually we'll explore providing default flag values on Ravenwood,
but let's get something cleaner available now.

Bug: 318841620
Test: ./frameworks/base/ravenwood/run-ravenwood-tests.sh
Change-Id: Ie6dadb259d3a35ee860d96f1b9a2e45f72002583