CVE-2017-13194: Fix poc for CVE-2017-13194
In poc enable_selective_overload flag is set to ENABLE_ALL before
memory allocation but it is getting unset before free() call, due
to which test crashes.
In this test CHECK_OVERFLOW is defined due to which while allocating
the memory it's access is disabled. As the enable_selective_overload
is unset before the free() call, free() will try to access the memory
without enabling memory access due to which test crashes.

Bug: 283706446
Test: Build and run the test

Change-Id: I84bafe72ce7341c14fb3f210939134c652f34923
Merged-In: I84bafe72ce7341c14fb3f210939134c652f34923