correct sdk_version
This change correct sdk_version for libraries. Before the sdk_version is
not set, and by default the library will be built against the
implementation of the framework. These two libraries do not need to
build against the implementation of the framework. Set the sdk_version
to the right scope.

Test: presubmit
Bug: 320485496
Change-Id: Ib49c2604a12254f81353a3469e687a340dee21a3