Seek audio to the video sync frame in tunnel-mode seek tests
MKV container formats don't have audio sync positions, so
SEEK_TO_PREVIOUS_SYNC doesn't actually sync to the previous sync
position that is associated with the video frame. Instead seek audio
exactly to the same location as the video extractor, this guarantees
that the seek position is earlier than the desired position, which
allows for the first video frame to be rendered. Without this change,
the audio seek will be to the audio position AFTER the desired seek
time, which means some tunnel mode implementations will drop the first
video frame, and cause the test to fail.

Note: The MKV container extractor should still be updated with a
proper implementation of SEEK_TO_PREVIOUS_SYNC, which will be tracked
separately.

Bug: 304712545
Test: atest android.media.decoder.cts.DecodeOnlyTest#testTunneledPerfectSeekInitialPeekOffHevc
Change-Id: I9ff9eec2a71d5102094392d195370077bf10d476