add logging around telecom-app binding + increase bind time
A cluster of CUJ test failures came in on March 4th. Upon inspection,
the test apps appear to be taking an usual amount of time to bind in
some cases (over 10 seconds in some cases). It's unclear what is causing
the bind process to delay since there are a lack in logs for when the
bind request is sent.

To better understand the binding delay and cirlce in on the root cause,
some logs have been added surrounding the bind process. Additonally, the
bind timeout has increased from 5 seconds to 10 seconds.

Bug: 328092080
Test: atest CtsTelecomCujTestCases
Change-Id: I733f10cbc1058bfe0cbf1fe47c47e59099568d95