Remove go to browser instruction for Instant App test
It is a SHOULD (optional) feature so shouldn't be in CTS. See
https://source.android.com/docs/compatibility/14/android-14-cdd#315_instant_apps
[C-1-6]:
For Instant Apps launched via web intents, as defined by using an intent
with action set to Intent.ACTION_VIEW and with a scheme of "http" or
"https", an additional user affordance SHOULD allow the user not to
launch the Instant App and launch the associated link with the
configured web browser, if a browser is available on the device.

Also update the manifest with the new requirement ids.

Bug: 327993911
Test: m CtsVerifier
      m CtsVerifierInstantApp
      adb install -r -g CtsVerifier.apk
      adb install -r --instant CtsVerifierInstantApp.apk
      Launch CtsVerifier, Instant Apps Notification Test
Change-Id: Idc3442fcd318518d071c884d667ecec13f39f1b3