Fixes the SplitAttributesCalculatorTest flakiness
The test starts the activity again in order to exit pip, but that
also makes the TFOrganizer applying a WCT to create a new TF to embed
the launching Activity. A new transition was created because of the
WCT and therefore queued up the later exit-pip transition (so the
pip activity was stayed on the pip task).

This CL starts the activity from the pip activity to avoid TFOrganizer
applying WCT to create new TF.

Bug: 316834237
Test: verify testSplitAttributesCalculatorInvocation_pip over 100 times
Change-Id: I8e3e577ebf8bb4f967466a183b9d20ed68020e97