Disable some CtsPackageInstallTestCases tests given FRP enforcement
Some existing tests rely on the fact that the user can override the FRP
state from shell.  This is one of the way an attacker can use to bypass
FRP, and is now prevented by the FRP hardening work.

Disable the test when the feature flag is on. The ground truth of the
FRP state can't be change from shell, not on user build.

Bug: 303188712
Test: atest both test cases (skipped when the flag is on)
Change-Id: Id48edf90c0931c4c1848cf5000a45ac69a7cd6a5