Don't ensure bar contrast when bar is tap through
The framework won't add a scrim for the gesture navigation bar.

This CL would also opt-out the edge-to-edge enforcement because we are
going to disable the scrim for status bar for apps targeting 35+ since
V. The test about status bar will be moved to a lower target SDK level
later.

Fix: 329140103
Bug: 309578419
Test: atest EnsureBarContrastTest
Change-Id: I66819b543c999e0a65737bc74037eaf7a6057f74