[CTS] [MTE] Reduce duplication in tests.
In particular, the TaggingTest class is common for the
manifest_disabled, manifest_enabled_sdk_29, manifest_enabled_sdk_30, and
sdk_29 apps. Move them to the common folder, and adjust the hostside
driver a little to use the generic class.

The sdk_30 and sdk_30_memtag cannot get this treatment as their
TaggingTest is unique, and both apps need to be installed at the same
time as they cross-communicate.

Also fixes up some spurious whitespace and upload warnings.

Bug: N/A
Test: atest CtsTaggingHostTestCases
Change-Id: If19f605164af0ff7a040cf5076bee25510c18694