Handle new CallStyle action layout
The new CallStyle action layout requires the icon and text to be
centered together in the action button, which I implemented using an
ImageSpan.

Unfortunately, this means that the text of the Answer button is no
longer "Answer", but "\ufffd\u00a0Answer", which causes
require/findViewWithText to fail.

Therefore, add require/findViewWithTextContaining, and use those instead
to verify that the action buttons are displaying their labels.

Bug: 268733030
Bug: 329378435
Fixes: 325494406
Test: NotificationTemplateTest
Merged-In: If1f7d591f96e615025739c3b8dd430c6669eb99c
Change-Id: If1f7d591f96e615025739c3b8dd430c6669eb99c