Update CTS test for Android Security b/272024837
On new Android versions (V and above), when the PendingIntent received by MediaSession#setMediaButtonReceiver() points to an activity, an exception is thrown instead of ignoring the call. The existing test implementation does not handle this exception, resulting in a failure.

This commit updates the test to handle any possible exceptions that may occur while setting the media button receiver.

Bug: 298292855
Test: Ran the new testcase on main with/without patch

Change-Id: Iabb677defb87242f3c7828064de147a095bacb59