Enable VP9 profile tests
Enable the value pair of profile and level information for VP9 clips.
Some VP9 clips may not contain the level information in CSD or may not
have a CSD at all. In such cases, profile can be synthesized from the
first frame, but level cannot be. So, this change adds a check to verify
the level only if the CSD contains it.

Test: android.mediav2.cts.ExtractorTest
Bug: 159582475
Bug: 295804596

Change-Id: I62d5849634d11b0a4b72baecea4833681440dba8