Create test marker at iteration level
This moves the creation of the test markers for the two EditTexts in
testInputConnectionWhenAddAndRemoveAltFocusableImFlagInFocus at the
iteration level, which attemps to fix the flakiness of the test.

Additionally specifies an EditText tag when creating the test marker, to
tell them appart in tests with multiple edit texts, or multiple
iterations.

Bug: 328861282
Test: atest FocusHandlingTest
Change-Id: If3169dcd59f61d72769277ebf6c6640bf052cced