webkit: Replace sendPointerSync with tap helper
None of our tests use sendPointerSync for anything other than doing a tap event.
We recently had an issue where one of our tests was not doing the up event which led to an issue with focus.

Adding a general tap event instead so that this issue can't happen again and removing direct access to sendPointerSync.

Bug: 330290472
Fixes: 330290472
Change-Id: Ie3d4d1d35f68e433f3837292f9c4ad9d8e9f1876