Camera: Fix flakey LogicalCameraDeviceTest
In Java, garbage collection isn't bound by syntactic block, but rather
by last use.

For an ImageReader, even if its surface is used by the camera, if we
don't reference to it at the time of camera usage, garbage collection
may close the ImageReader.

Do an explicit reference at the end to keep the ImageReader alive.

Test: atest LogicalCameraDeviceTest#testBasicPhysicalStreaming
Bug: 329360235
Change-Id: Ief05db5699edbf3d8f14ca05df8bda0a7f2ba7cf