Allowlist content provider for sandbox testing
SDK sandboxes cannot access ContentProviders unless specifically
allowlisted. In order to allow the test to be run in SDK sandbox mode,
allowlist the required test CPs before running the test.

Additionally, the test ContentProvider needs to be exported so that the
sandbox (that has a different package name from the test app) can access
it. Modify an existing test that checks whether this test CP is exported
or not to check a different CP not used by the sandbox instead.

Bug: 300435558
Test: atest ContentResolverTest  -- --enable-optional-parameterization --enable-parameterized-modules --module-parameter run_on_sdk_sandbox
Test: atest ContentQueryMapTest  -- --enable-optional-parameterization --enable-parameterized-modules --module-parameter run_on_sdk_sandbox
Test: atest PackageManagerTest
Change-Id: I74ba1725e09b5755eb36c07e3cbf0548a1a9d7ea