[2/n] Disable cts tests with freeform activities when desktop windowing enabled
Desktop Windowing is currently only available on selected devices when the
desktop windowing flag is enabled. Desktop windowing currently changes the
behaviour of freeform activities and thus when this feature is available
the legacy behaviour of freeform activities will not be respected and any
tests that depend on this behaviour will fail. So we ignore these tests
for tablets if desktop windowing is enabled while we work to have the
legacy freeform behaviour and desktop mode coexist.

Bug: 330152508
Bug: 331170652
Test: atest CtsWindowManagerDeviceInsets:WindowInsetsAnimationTests
atest CtsWindowManagerDeviceInsets:WindowInsetsControllerTests
atest CtsWindowManagerDeviceDisplay:MultiDisplayLockedKeyguardTests
atest CtsWindowManagerDeviceDisplay:FreeformWindowingModeTests
atest CtsWindowManagerDeviceActivity:ActivityLifecycleFreeformTests

Change-Id: Ib293bbbbda2186febf845e8e964da366f721452d