[DO NOT MERGE] Adjust Toast test for Automotive
In Android S the toast functionality was expanded to show an application
icon next to the text. For Automotive, this does not work out of the box
with the AOSP CarSystemUI, most likely due to the headless user 0 setup
in Automotive.
When requesting a toast on CarSystemUI, based on s-car-lts-fs-release
(SAG3.230621.002), SystemUIToast prints "user hasn't been fully
initialized, not showing an app icon for packageName=...".
This has since been fixed in Android T, so that the Toast icon will work
for Automotive as well out of the box:
    ec9a21c86b1b2ee183697276fff92f96a0cca555
    Toast app icons should show for secondary users

This patch adjusts the instructions for Automotive to not look for an
icon in the toast.

Bug: 331174647
Test: manual "Toast test" in CtsVerifier

Change-Id: I4027b54412509dfdadb3c8bfb7f4b0508138c74d