[sb] Remove the color-counting assertion in LightBarTests
This check is asserting that there are less than 5% of pixels that have
an unaccounted for color. As far as I can tell, there is no assertion in
CDD that requires the status bar icons to be monochrome. So it seems
okay to remove this check in the tests.

Test: atest LightBarTests
Fixes: 330965837
Bug: 330707481
Flag: ACONFIG com.android.settingslib.flags.new_status_bar_icons TEAMFOOD
Change-Id: I3bea48ce39bff523e925d063966c31595997fbee