Cleanup unncessary version check
We have tests in SdkAvailabilityTest to verify that the extensions
version is the latest, so there is no need to test for older version
which are stable (we still need similar checks for future extensions
features).

Also move some @Ignored test cases to @Flaky to monitor if the failure
is still happening.

Bug: 306666082
Bug: 332251465
Test: pass presubmit
Change-Id: I333b3d505d79c3fa8b12917841cc2854a1c9b909