Make SQLite CTS tests hermetic
CTS tests should be hermetic, which means the final state should be
the same as the initial state, as much as possible.  Some of the
database tests failed to close their database, which violated the
assumptions of other tests.

It would be possible to remove the "clean system" assumption from the
failing tests, but it is good practice for all tests to clean up after
themselves.

This was tested more thoroughly with a custom image that reported when
any database test started or ended with an already-open database.

Test: atest
 * CtsDatabaseTestCases

Bug: 325676267
Change-Id: Ia90c843f78431f852e0f8cefee38831b2b60df9c